rag/requirements.txt
# Core dependencies
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.10.3

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Gradio for chatbot UI
gradio==5.44.1

# LangChain and LLM dependencies
langchain==0.3.27
langchain-ollama==0.3.7
langchain-community==0.3.29

# Vector database (Qdrant - persistent, fast, reliable)
qdrant-client==1.15.1

# PDF processing
pypdf==6.0.0

# Data processing
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6

# Environment management
python-dotenv==1.1.0
